restructuring pagination components as per review comments
