+++ vehicles/src/common/decorator/api-paginate-response.ts
@@ -1,6 +1,6 @@
 import { applyDecorators, Type } from '@nestjs/common';
 import { ApiExtraModels, ApiOkResponse, getSchemaPath } from '@nestjs/swagger';
-import { PaginationDto } from './pagination';
+import { PaginationDto } from '../dto/paginate/pagination';
 
 export const ApiPaginatedResponse = <TModel extends Type<any>>(
   model: TModel,
