+++ vehicles/src/common/interface/pagination.interface.ts
@@ -1,4 +1,4 @@
-import { PageOptionsDto } from '../class/page-options';
+import { PageOptionsDto } from '../dto/paginate/page-options';
 
 export interface PageMetaDtoParameters {
   pageOptionsDto: PageOptionsDto;
